+++ app/Http/Controllers/v1/FolderApiController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers\v1;
 
 use App\Http\Controllers\Controller;
 use App\Models\Folder;
+use App\Models\Space;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 class FolderApiController extends Controller
 {
@@ -27,19 +29,52 @@ class FolderApiController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
-        $folder = Folder::query()->create($request->all());
-        return response()->json($folder, 201);
+        $request->validate([
+            'name' => 'required',
+            'parent_folder_id' => 'nullable|exists:folders,id',
+            'space_id' => 'required|exists:spaces,id',
+        ]);
+
+        $user = $request->user();
+
+        if (isset($request['parent_folder_id'])) {
+            $parentFolder = Folder::query()->find($request['parent_folder_id']);
+
+            if ($user->cannot('uploadInto', $parentFolder)) {
+                return response()->json(['message' => 'You do not have the right to create folders in this folder.'], Response::HTTP_FORBIDDEN);
+            }
+        } else {
+            $space = Space::query()->find($request['space_id']);
+
+            if ($user->cannot('uploadInto', $space)) {
+                return response()->json(['message' => 'You do not have the right to create folders in this space.'], Response::HTTP_FORBIDDEN);
+            }
+        }
+
+        $folder = Folder::create([
+            'name' => $request['name'],
+            'parent_folder_id' => $request['parent_folder_id'],
+            'space_id' => $request['space_id'],
+            'owner_id' => $user->id,
+            'is_shortcut' => $request['is_shortcut'] ?? false,
+            'shortcut_target_id' => $request['shortcut_target_id'] ?? null,
+        ]);
+
+        return response()->json($folder, Response::HTTP_CREATED);
     }
 
     /**
      * Display the specified resource.
      *
+     * @param Request $request
      * @param Folder $folder
      * @return JsonResponse
      */
-    public function show(Folder $folder): JsonResponse
+    public function show(Request $request, Folder $folder): JsonResponse
     {
-        $folder = Folder::query()->findOrFail($folder->id);
+        if ($request->user()->cannot('view', $folder)) {
+            return response()->json(['message' => 'You do not have the required privileges to access to this folder.'], Response::HTTP_FORBIDDEN);
+        }
         return response()->json($folder);
     }
 
@@ -52,7 +87,10 @@ class FolderApiController extends Controller
      */
     public function update(Request $request, Folder $folder): JsonResponse
     {
-        $folder = Folder::query()->findOrFail($folder->id);
+        if ($request->user()->cannot('edit', $folder)) {
+            return response()->json(['message' => 'You do not have the required privileges to edit this folder.'], Response::HTTP_FORBIDDEN);
+        }
+
         $folder->update($request->all());
         return response()->json($folder);
     }
@@ -60,12 +98,15 @@ class FolderApiController extends Controller
     /**
      * Remove the specified resource from storage.
      *
+     * @param Request $request
      * @param Folder $folder
      * @return JsonResponse
      */
-    public function destroy(Folder $folder): JsonResponse
+    public function destroy(Request $request, Folder $folder): JsonResponse
     {
-        $folder = Folder::query()->findOrFail($folder->id);
+        if ($request->user()->cannot('edit', $folder)) {
+            return response()->json(['message' => 'You do not have the required privileges to delete this folder.'], Response::HTTP_FORBIDDEN);
+        }
         $folder->delete();
         return response()->json(null, 204);
     }
